Renamed constants/constants.js to constants/menu.js
